Day 8 Start
Changed grid size. now grid size and vision is dynamic. the creatures do actually work a lot better. it makes a lot of sense. however, the classical version gets stuck going in a small circle because it thinks there's a wall right next to it but it's actually far. thinking of normalizing the vision. will see.